moved align, bwt, and mash into new search package. (#438)
* moved align, bwt, and mash into new search package.

* updated import paths.
