+++ search/align/align_test.go
@@ -3,9 +3,9 @@ package align_test
 import (
 	"testing"
 
-	"github.com/bebop/poly/align"
-	"github.com/bebop/poly/align/matrix"
 	"github.com/bebop/poly/alphabet"
+	"github.com/bebop/poly/search/align"
+	"github.com/bebop/poly/search/align/matrix"
 )
 
 func TestNeedlemanWunsch(t *testing.T) {
