+++ search/bwt/bitvector.go
