+++ search/bwt/bitvector_test.go
