+++ search/bwt/bwt.go
