+++ search/bwt/bwt_test.go
