+++ search/bwt/rsa_bitvector.go
