+++ search/bwt/rsa_bitvector_test.go
