+++ search/bwt/wavelet.go
