+++ search/bwt/wavelet_test.go
